ocr/requirements.txt
# Dependencies for the OCR container.

fastapi==0.115.12
uvicorn[standard]==0.34.2

pyspellchecker

git+https://github.com/facebookresearch/detectron2.git@v0.4#egg=detectron2
pillow==9.5.0

python-doctr
layoutparser
